Adjust withdraw tests
