+++ packages/app/src/features/dialogs/savings/withdraw/e2e/base/usds-account/WithdrawMaxUSDC.test-e2e.ts
@@ -1,0 +1,70 @@
+import { SavingsPageObject } from '@/pages/Savings.PageObject'
+import { BASE_DEFAULT_BLOCK_NUMBER } from '@/test/e2e/constants'
+import { setup } from '@/test/e2e/setup'
+import { test } from '@playwright/test'
+import { base } from 'viem/chains'
+import { SavingsDialogPageObject } from '../../../../common/e2e/SavingsDialog.PageObject'
+
+test.describe('Withdraw Max USDC', () => {
+  let savingsPage: SavingsPageObject
+  let withdrawDialog: SavingsDialogPageObject
+
+  test.beforeEach(async ({ page }) => {
+    const testContext = await setup(page, {
+      blockchain: {
+        chainId: base.id,
+        blockNumber: BASE_DEFAULT_BLOCK_NUMBER,
+      },
+      initialPage: 'savings',
+      account: {
+        type: 'connected-random',
+        assetBalances: {
+          ETH: 1,
+          sUSDS: 10_000,
+        },
+      },
+    })
+
+    savingsPage = new SavingsPageObject(testContext)
+
+    await savingsPage.clickWithdrawFromAccountButtonAction()
+    withdrawDialog = new SavingsDialogPageObject({ testContext, type: 'withdraw' })
+    await withdrawDialog.selectAssetAction('USDC')
+    await withdrawDialog.clickMaxAmountAction()
+  })
+
+  test('has correct action plan', async () => {
+    await withdrawDialog.actionsContainer.expectActions([
+      { type: 'approve', asset: 'sUSDS' },
+      { type: 'withdrawFromSavings', asset: 'USDC', savingsAsset: 'sUSDS', mode: 'withdraw' },
+    ])
+  })
+
+  test('displays transaction overview', async () => {
+    await withdrawDialog.expectNativeRouteTransactionOverview({
+      routeItems: [
+        {
+          tokenAmount: '10,000.00 sUSDS',
+          tokenUsdValue: '$10,128.65',
+        },
+        {
+          tokenAmount: '10,128.65 USDS',
+          tokenUsdValue: '$10,128.65',
+        },
+      ],
+      outcome: '10,128.65 USDC',
+      outcomeUsd: '$10,128.65',
+    })
+
+    await withdrawDialog.expectUpgradeSwitchToBeHidden()
+  })
+
+  test('executes withdraw', async () => {
+    await withdrawDialog.actionsContainer.acceptAllActionsAction(2)
+
+    await withdrawDialog.expectSuccessPage()
+    await withdrawDialog.clickBackToSavingsButton()
+
+    await savingsPage.expectSupportedStablecoinBalance('USDC', '10,128.65')
+  })
+})
