+++ packages/app/src/features/dialogs/savings/withdraw/e2e/base/usds-account/WithdrawUSDS.test-e2e.ts
@@ -1,0 +1,70 @@
+import { SavingsPageObject } from '@/pages/Savings.PageObject'
+import { BASE_DEFAULT_BLOCK_NUMBER } from '@/test/e2e/constants'
+import { setup } from '@/test/e2e/setup'
+import { test } from '@playwright/test'
+import { base } from 'viem/chains'
+import { SavingsDialogPageObject } from '../../../../common/e2e/SavingsDialog.PageObject'
+
+test.describe('Withdraw USDS', () => {
+  let savingsPage: SavingsPageObject
+  let withdrawDialog: SavingsDialogPageObject
+
+  test.beforeEach(async ({ page }) => {
+    const testContext = await setup(page, {
+      blockchain: {
+        chainId: base.id,
+        blockNumber: BASE_DEFAULT_BLOCK_NUMBER,
+      },
+      initialPage: 'savings',
+      account: {
+        type: 'connected-random',
+        assetBalances: {
+          ETH: 1,
+          sUSDS: 10_000,
+        },
+      },
+    })
+
+    savingsPage = new SavingsPageObject(testContext)
+
+    await savingsPage.clickWithdrawFromAccountButtonAction()
+    withdrawDialog = new SavingsDialogPageObject({ testContext, type: 'withdraw' })
+    await withdrawDialog.fillAmountAction(1000)
+  })
+
+  test('has correct action plan', async () => {
+    await withdrawDialog.actionsContainer.expectActions([
+      { type: 'approve', asset: 'sUSDS' },
+      { type: 'withdrawFromSavings', asset: 'USDS', savingsAsset: 'sUSDS', mode: 'withdraw' },
+    ])
+  })
+
+  test('displays transaction overview', async () => {
+    await withdrawDialog.expectNativeRouteTransactionOverview({
+      routeItems: [
+        {
+          tokenAmount: '987.30 sUSDS',
+          tokenUsdValue: '$1,000.00',
+        },
+        {
+          tokenAmount: '1,000.00 USDS',
+          tokenUsdValue: '$1,000.00',
+        },
+      ],
+      outcome: '1,000.00 USDS',
+      outcomeUsd: '$1,000.00',
+    })
+
+    await withdrawDialog.expectUpgradeSwitchToBeHidden()
+  })
+
+  test('executes withdraw', async () => {
+    await withdrawDialog.actionsContainer.acceptAllActionsAction(2)
+
+    await withdrawDialog.expectSuccessPage()
+    await withdrawDialog.clickBackToSavingsButton()
+
+    await savingsPage.expectSavingsAccountBalance({ balance: '9,012.70', estimatedValue: '9,128.654892' })
+    await savingsPage.expectSupportedStablecoinBalance('USDS', '1,000')
+  })
+})
