--- packages/app/src/features/dialogs/savings/withdraw/e2e/mainnet/SendUSDS.test-e2e.ts
@@ -1,171 +1,0 @@
-import { SavingsPageObject } from '@/pages/Savings.PageObject'
-import { DEFAULT_BLOCK_NUMBER, TOKENS_ON_FORK } from '@/test/e2e/constants'
-import { setup } from '@/test/e2e/setup'
-import { randomAddress } from '@/test/utils/addressUtils'
-import { test } from '@playwright/test'
-import { mainnet } from 'viem/chains'
-import { SavingsDialogPageObject } from '../../../common/e2e/SavingsDialog.PageObject'
-
-test.describe('Send USDS (withdrawing from sUSDS)', () => {
-  let savingsPage: SavingsPageObject
-  let sendDialog: SavingsDialogPageObject
-  const receiver = randomAddress('bob')
-  const amount = 7000
-  const usds = TOKENS_ON_FORK[mainnet.id].USDS
-
-  test.beforeEach(async ({ page }) => {
-    const testContext = await setup(page, {
-      blockchain: {
-        chainId: mainnet.id,
-        blockNumber: DEFAULT_BLOCK_NUMBER,
-      },
-      initialPage: 'savings',
-      account: {
-        type: 'connected-random',
-        assetBalances: {
-          ETH: 1,
-          USDS: 10_000,
-        },
-      },
-    })
-
-    savingsPage = new SavingsPageObject(testContext)
-
-    await savingsPage.clickDepositButtonAction('USDS')
-    const depositDialog = new SavingsDialogPageObject({ testContext, type: 'deposit' })
-    await depositDialog.fillAmountAction(10_000)
-    await depositDialog.actionsContainer.acceptAllActionsAction(2)
-    await depositDialog.clickBackToSavingsButton()
-
-    await savingsPage.clickSendFromAccountButtonAction()
-    sendDialog = new SavingsDialogPageObject({ testContext, type: 'send' })
-    await sendDialog.fillAmountAction(amount)
-    await sendDialog.fillReceiverAction(receiver)
-  })
-
-  test('uses native sUSDS withdraw and send', async () => {
-    await sendDialog.actionsContainer.expectActions([
-      { type: 'withdrawFromSavings', asset: 'USDS', savingsAsset: 'sUSDS', mode: 'send' },
-    ])
-  })
-
-  test('displays transaction overview', async () => {
-    await sendDialog.expectNativeRouteTransactionOverview({
-      routeItems: [
-        {
-          tokenAmount: '6,881.24 sUSDS',
-          tokenUsdValue: '$7,000.00',
-        },
-        {
-          tokenAmount: '7,000.00 USDS',
-          tokenUsdValue: '$7,000.00',
-        },
-      ],
-      outcome: '7,000.00 USDS',
-      outcomeUsd: '$7,000.00',
-    })
-  })
-
-  test('executes send', async () => {
-    await sendDialog.expectReceiverTokenBalance({
-      receiver,
-      token: usds,
-      expectedBalance: 0,
-    })
-
-    await sendDialog.actionsContainer.acceptAllActionsAction(1)
-    await sendDialog.expectSuccessPage()
-
-    await sendDialog.expectReceiverTokenBalance({
-      receiver,
-      token: usds,
-      expectedBalance: amount,
-    })
-
-    await sendDialog.clickBackToSavingsButton()
-    await savingsPage.expectSavingsAccountBalance({ balance: '2,949.10', estimatedValue: '3,000.0001867' })
-    await savingsPage.expectSupportedStablecoinBalance('USDS', '-')
-  })
-})
-
-test.describe('Send USDS (withdrawing from sDAI)', () => {
-  let savingsPage: SavingsPageObject
-  let sendDialog: SavingsDialogPageObject
-  const receiver = randomAddress('bob')
-  const amount = 7000
-  const usds = TOKENS_ON_FORK[mainnet.id].USDS
-
-  test.beforeEach(async ({ page }) => {
-    const testContext = await setup(page, {
-      blockchain: {
-        chainId: mainnet.id,
-        blockNumber: DEFAULT_BLOCK_NUMBER,
-      },
-      initialPage: 'savings',
-      account: {
-        type: 'connected-random',
-        assetBalances: {
-          ETH: 1,
-          sDAI: 10_000,
-        },
-      },
-    })
-
-    savingsPage = new SavingsPageObject(testContext)
-
-    await savingsPage.clickSendFromAccountButtonAction()
-    sendDialog = new SavingsDialogPageObject({ testContext, type: 'send' })
-    await sendDialog.selectAssetAction('USDS')
-    await sendDialog.fillAmountAction(amount)
-    await sendDialog.fillReceiverAction(receiver)
-  })
-
-  test('uses migrate sDAI to USDS action', async () => {
-    await sendDialog.actionsContainer.expectActions([
-      { type: 'approve', asset: 'sDAI' },
-      { type: 'withdrawFromSavings', asset: 'USDS', savingsAsset: 'sDAI', mode: 'send' },
-    ])
-  })
-
-  test('displays transaction overview', async () => {
-    await sendDialog.expectNativeRouteTransactionOverview({
-      routeItems: [
-        {
-          tokenAmount: '6,218.91 sDAI',
-          tokenUsdValue: '$7,000.00',
-        },
-        {
-          tokenAmount: '7,000.00 DAI',
-          tokenUsdValue: '$7,000.00',
-        },
-        {
-          tokenAmount: '7,000.00 USDS',
-          tokenUsdValue: '$7,000.00',
-        },
-      ],
-      outcome: '7,000.00 USDS',
-      outcomeUsd: '$7,000.00',
-    })
-  })
-
-  test('executes send', async () => {
-    await sendDialog.expectReceiverTokenBalance({
-      receiver,
-      token: usds,
-      expectedBalance: 0,
-    })
-
-    await sendDialog.actionsContainer.acceptAllActionsAction(2)
-    await sendDialog.expectSuccessPage()
-
-    await sendDialog.expectReceiverTokenBalance({
-      receiver,
-      token: usds,
-      expectedBalance: amount,
-    })
-
-    await sendDialog.clickBackToSavingsButton()
-    await savingsPage.expectSavingsAccountBalance({ balance: '3,781.09', estimatedValue: '4,255.9920127' })
-    await savingsPage.expectSupportedStablecoinBalance('USDS', '-')
-  })
-})
