+++ packages/app/src/features/dialogs/savings/withdraw/e2e/mainnet/dai-account/SendUSDC.test-e2e.ts
@@ -1,40 +1,39 @@
+import { SavingsDialogPageObject } from '@/features/dialogs/savings/common/e2e/SavingsDialog.PageObject'
 import { SavingsPageObject } from '@/pages/Savings.PageObject'
-import { BASE_DEFAULT_BLOCK_NUMBER, TOKENS_ON_FORK } from '@/test/e2e/constants'
+import { DEFAULT_BLOCK_NUMBER, TOKENS_ON_FORK } from '@/test/e2e/constants'
 import { setup } from '@/test/e2e/setup'
 import { randomAddress } from '@/test/utils/addressUtils'
 import { test } from '@playwright/test'
-import { base } from 'viem/chains'
-import { SavingsDialogPageObject } from '../../../common/e2e/SavingsDialog.PageObject'
+import { mainnet } from 'viem/chains'
 
 test.describe('Send USDC', () => {
   let savingsPage: SavingsPageObject
   let sendDialog: SavingsDialogPageObject
-
   const receiver = randomAddress('bob')
   const amount = 7000
-  const usdc = TOKENS_ON_FORK[base.id].USDC
+  const usdc = TOKENS_ON_FORK[mainnet.id].USDC
 
   test.beforeEach(async ({ page }) => {
     const testContext = await setup(page, {
       blockchain: {
-        chainId: base.id,
-        blockNumber: BASE_DEFAULT_BLOCK_NUMBER,
+        chainId: mainnet.id,
+        blockNumber: DEFAULT_BLOCK_NUMBER,
       },
       initialPage: 'savings',
       account: {
         type: 'connected-random',
         assetBalances: {
           ETH: 1,
-          sUSDS: 10_000,
+          sDAI: 10_000,
         },
       },
     })
 
     savingsPage = new SavingsPageObject(testContext)
-
+    await savingsPage.clickSavingsNavigationItemAction('DAI')
     await savingsPage.clickSendFromAccountButtonAction()
-    sendDialog = new SavingsDialogPageObject({ testContext, type: 'send' })
 
+    sendDialog = new SavingsDialogPageObject({ testContext, type: 'send' })
     await sendDialog.selectAssetAction('USDC')
     await sendDialog.fillAmountAction(amount)
     await sendDialog.fillReceiverAction(receiver)
@@ -42,8 +41,8 @@ test.describe('Send USDC', () => {
 
   test('has correct action plan', async () => {
     await sendDialog.actionsContainer.expectActions([
-      { type: 'approve', asset: 'sUSDS' },
-      { type: 'withdrawFromSavings', asset: 'USDC', savingsAsset: 'sUSDS', mode: 'send' },
+      { type: 'approve', asset: 'sDAI' },
+      { type: 'withdrawFromSavings', asset: 'USDC', savingsAsset: 'sDAI', mode: 'send' },
     ])
   })
 
@@ -51,11 +50,11 @@ test.describe('Send USDC', () => {
     await sendDialog.expectNativeRouteTransactionOverview({
       routeItems: [
         {
-          tokenAmount: '6,911.09 sUSDS',
+          tokenAmount: '6,218.91 sDAI',
           tokenUsdValue: '$7,000.00',
         },
         {
-          tokenAmount: '7,000.00 USDS',
+          tokenAmount: '7,000.00 DAI',
           tokenUsdValue: '$7,000.00',
         },
         {
@@ -85,7 +84,7 @@ test.describe('Send USDC', () => {
     })
 
     await sendDialog.clickBackToSavingsButton()
-    await savingsPage.expectSavingsAccountBalance({ balance: '3,088.91', estimatedValue: '3,128.6548910' })
+    await savingsPage.expectSavingsAccountBalance({ balance: '3,781.09', estimatedValue: '4,255.9920127' })
     await savingsPage.expectSupportedStablecoinBalance('USDC', '-')
   })
 })
