+++ packages/app/src/features/dialogs/savings/withdraw/e2e/mainnet/usds-account/SendUSDC.test-e2e.ts
@@ -1,12 +1,12 @@
+import { SavingsDialogPageObject } from '@/features/dialogs/savings/common/e2e/SavingsDialog.PageObject'
 import { SavingsPageObject } from '@/pages/Savings.PageObject'
 import { DEFAULT_BLOCK_NUMBER, TOKENS_ON_FORK } from '@/test/e2e/constants'
 import { setup } from '@/test/e2e/setup'
 import { randomAddress } from '@/test/utils/addressUtils'
 import { test } from '@playwright/test'
 import { mainnet } from 'viem/chains'
-import { SavingsDialogPageObject } from '../../../common/e2e/SavingsDialog.PageObject'
 
-test.describe('Send USDC (withdrawing from sUSDS)', () => {
+test.describe('Send USDC', () => {
   let savingsPage: SavingsPageObject
   let sendDialog: SavingsDialogPageObject
   const receiver = randomAddress('bob')
@@ -37,6 +37,7 @@ test.describe('Send USDC (withdrawing from sUSDS)', () => {
     await depositDialog.actionsContainer.acceptAllActionsAction(2)
     await depositDialog.clickBackToSavingsButton()
 
+    await savingsPage.clickSavingsNavigationItemAction('USDS')
     await savingsPage.clickSendFromAccountButtonAction()
     sendDialog = new SavingsDialogPageObject({ testContext, type: 'send' })
     await sendDialog.selectAssetAction('USDC')
@@ -90,88 +91,6 @@ test.describe('Send USDC (withdrawing from sUSDS)', () => {
 
     await sendDialog.clickBackToSavingsButton()
     await savingsPage.expectSavingsAccountBalance({ balance: '2,949.10', estimatedValue: '3,000.0003735' })
-    await savingsPage.expectSupportedStablecoinBalance('USDC', '-')
-  })
-})
-
-test.describe('Send USDC (withdrawing from sDAI)', () => {
-  let savingsPage: SavingsPageObject
-  let sendDialog: SavingsDialogPageObject
-  const receiver = randomAddress('bob')
-  const amount = 7000
-  const usdc = TOKENS_ON_FORK[mainnet.id].USDC
-
-  test.beforeEach(async ({ page }) => {
-    const testContext = await setup(page, {
-      blockchain: {
-        chainId: mainnet.id,
-        blockNumber: DEFAULT_BLOCK_NUMBER,
-      },
-      initialPage: 'savings',
-      account: {
-        type: 'connected-random',
-        assetBalances: {
-          ETH: 1,
-          sDAI: 10_000,
-        },
-      },
-    })
-
-    savingsPage = new SavingsPageObject(testContext)
-
-    await savingsPage.clickSendFromAccountButtonAction()
-    sendDialog = new SavingsDialogPageObject({ testContext, type: 'send' })
-    await sendDialog.selectAssetAction('USDC')
-    await sendDialog.fillAmountAction(amount)
-    await sendDialog.fillReceiverAction(receiver)
-  })
-
-  test('uses convert sDAI to USDS action', async () => {
-    await sendDialog.actionsContainer.expectActions([
-      { type: 'approve', asset: 'sDAI' },
-      { type: 'withdrawFromSavings', asset: 'USDC', savingsAsset: 'sDAI', mode: 'send' },
-    ])
-  })
-
-  test('displays transaction overview', async () => {
-    await sendDialog.expectNativeRouteTransactionOverview({
-      routeItems: [
-        {
-          tokenAmount: '6,218.91 sDAI',
-          tokenUsdValue: '$7,000.00',
-        },
-        {
-          tokenAmount: '7,000.00 DAI',
-          tokenUsdValue: '$7,000.00',
-        },
-        {
-          tokenAmount: '7,000.00 USDC',
-          tokenUsdValue: '$7,000.00',
-        },
-      ],
-      outcome: '7,000.00 USDC',
-      outcomeUsd: '$7,000.00',
-    })
-  })
-
-  test('executes send', async () => {
-    await sendDialog.expectReceiverTokenBalance({
-      receiver,
-      token: usdc,
-      expectedBalance: 0,
-    })
-
-    await sendDialog.actionsContainer.acceptAllActionsAction(2)
-    await sendDialog.expectSuccessPage()
-
-    await sendDialog.expectReceiverTokenBalance({
-      receiver,
-      token: usdc,
-      expectedBalance: amount,
-    })
-
-    await sendDialog.clickBackToSavingsButton()
-    await savingsPage.expectSavingsAccountBalance({ balance: '3,781.09', estimatedValue: '4,255.9920127' })
     await savingsPage.expectSupportedStablecoinBalance('USDC', '-')
   })
 })
