+++ packages/app/src/features/dialogs/savings/withdraw/e2e/mainnet/usds-account/SendUSDS.test-e2e.ts
@@ -1,17 +1,17 @@
+import { SavingsDialogPageObject } from '@/features/dialogs/savings/common/e2e/SavingsDialog.PageObject'
 import { SavingsPageObject } from '@/pages/Savings.PageObject'
 import { DEFAULT_BLOCK_NUMBER, TOKENS_ON_FORK } from '@/test/e2e/constants'
 import { setup } from '@/test/e2e/setup'
 import { randomAddress } from '@/test/utils/addressUtils'
 import { test } from '@playwright/test'
 import { mainnet } from 'viem/chains'
-import { SavingsDialogPageObject } from '../../../common/e2e/SavingsDialog.PageObject'
 
-test.describe('Send DAI on Mainnet', () => {
+test.describe('Send USDS', () => {
   let savingsPage: SavingsPageObject
   let sendDialog: SavingsDialogPageObject
   const receiver = randomAddress('bob')
   const amount = 7000
-  const dai = TOKENS_ON_FORK[mainnet.id].DAI
+  const usds = TOKENS_ON_FORK[mainnet.id].USDS
 
   test.beforeEach(async ({ page }) => {
     const testContext = await setup(page, {
@@ -24,22 +24,29 @@ test.describe('Send DAI on Mainnet', () => {
         type: 'connected-random',
         assetBalances: {
           ETH: 1,
-          sDAI: 10_000,
+          USDS: 10_000,
         },
       },
     })
 
     savingsPage = new SavingsPageObject(testContext)
-    await savingsPage.clickSendFromAccountButtonAction()
 
+    await savingsPage.clickDepositButtonAction('USDS')
+    const depositDialog = new SavingsDialogPageObject({ testContext, type: 'deposit' })
+    await depositDialog.fillAmountAction(10_000)
+    await depositDialog.actionsContainer.acceptAllActionsAction(2)
+    await depositDialog.clickBackToSavingsButton()
+
+    await savingsPage.clickSavingsNavigationItemAction('USDS')
+    await savingsPage.clickSendFromAccountButtonAction()
     sendDialog = new SavingsDialogPageObject({ testContext, type: 'send' })
     await sendDialog.fillAmountAction(amount)
     await sendDialog.fillReceiverAction(receiver)
   })
 
-  test('uses native sDai withdraw and send', async () => {
+  test('has correct action plan', async () => {
     await sendDialog.actionsContainer.expectActions([
-      { type: 'withdrawFromSavings', asset: 'DAI', savingsAsset: 'sDAI', mode: 'send' },
+      { type: 'withdrawFromSavings', asset: 'USDS', savingsAsset: 'sUSDS', mode: 'send' },
     ])
   })
 
@@ -47,15 +54,15 @@ test.describe('Send DAI on Mainnet', () => {
     await sendDialog.expectNativeRouteTransactionOverview({
       routeItems: [
         {
-          tokenAmount: '6,218.91 sDAI',
+          tokenAmount: '6,881.24 sUSDS',
           tokenUsdValue: '$7,000.00',
         },
         {
-          tokenAmount: '7,000.00 DAI',
+          tokenAmount: '7,000.00 USDS',
           tokenUsdValue: '$7,000.00',
         },
       ],
-      outcome: '7,000.00 DAI',
+      outcome: '7,000.00 USDS',
       outcomeUsd: '$7,000.00',
     })
   })
@@ -63,7 +70,7 @@ test.describe('Send DAI on Mainnet', () => {
   test('executes send', async () => {
     await sendDialog.expectReceiverTokenBalance({
       receiver,
-      token: dai,
+      token: usds,
       expectedBalance: 0,
     })
 
@@ -72,13 +79,12 @@ test.describe('Send DAI on Mainnet', () => {
 
     await sendDialog.expectReceiverTokenBalance({
       receiver,
-      token: dai,
+      token: usds,
       expectedBalance: amount,
     })
 
     await sendDialog.clickBackToSavingsButton()
-
-    await savingsPage.expectSavingsAccountBalance({ balance: '3,781.09', estimatedValue: '4,255.9918184' })
-    await savingsPage.expectSupportedStablecoinBalance('DAI', '-')
+    await savingsPage.expectSavingsAccountBalance({ balance: '2,949.10', estimatedValue: '3,000.0001867' })
+    await savingsPage.expectSupportedStablecoinBalance('USDS', '-')
   })
 })
