+++ packages/app/src/features/dialogs/savings/withdraw/e2e/mainnet/usds-account/WithdrawMaxUSDC.test-e2e.ts
@@ -1,11 +1,11 @@
+import { SavingsDialogPageObject } from '@/features/dialogs/savings/common/e2e/SavingsDialog.PageObject'
 import { SavingsPageObject } from '@/pages/Savings.PageObject'
 import { DEFAULT_BLOCK_NUMBER } from '@/test/e2e/constants'
 import { setup } from '@/test/e2e/setup'
 import { test } from '@playwright/test'
 import { mainnet } from 'viem/chains'
-import { SavingsDialogPageObject } from '../../../common/e2e/SavingsDialog.PageObject'
 
-test.describe('Withdraw USDC from sUSDS', () => {
+test.describe('Withdraw max USDC', () => {
   let savingsPage: SavingsPageObject
   let withdrawDialog: SavingsDialogPageObject
 
@@ -33,6 +33,7 @@ test.describe('Withdraw USDC from sUSDS', () => {
     await depositDialog.actionsContainer.acceptAllActionsAction(2)
     await depositDialog.clickBackToSavingsButton()
 
+    await savingsPage.clickSavingsNavigationItemAction('USDS')
     await savingsPage.clickWithdrawFromAccountButtonAction()
     withdrawDialog = new SavingsDialogPageObject({ testContext, type: 'withdraw' })
     await withdrawDialog.selectAssetAction('USDC')
@@ -76,73 +77,5 @@ test.describe('Withdraw USDC from sUSDS', () => {
     await withdrawDialog.clickBackToSavingsButton()
 
     await savingsPage.expectSupportedStablecoinBalance('USDC', '10,000')
-  })
-})
-
-test.describe('Withdraw USDC from sDAI', () => {
-  let savingsPage: SavingsPageObject
-  let withdrawDialog: SavingsDialogPageObject
-
-  test.beforeEach(async ({ page }) => {
-    const testContext = await setup(page, {
-      blockchain: {
-        chainId: mainnet.id,
-        blockNumber: DEFAULT_BLOCK_NUMBER,
-      },
-      initialPage: 'savings',
-      account: {
-        type: 'connected-random',
-        assetBalances: {
-          ETH: 1,
-          sDAI: 10_000,
-        },
-      },
-    })
-
-    savingsPage = new SavingsPageObject(testContext)
-
-    await savingsPage.clickWithdrawFromAccountButtonAction()
-    withdrawDialog = new SavingsDialogPageObject({ testContext, type: 'withdraw' })
-    await withdrawDialog.selectAssetAction('USDC')
-    await withdrawDialog.clickMaxAmountAction()
-  })
-
-  test('has correct action plan', async () => {
-    await withdrawDialog.actionsContainer.expectActions([
-      { type: 'approve', asset: 'sDAI' },
-      { type: 'withdrawFromSavings', asset: 'USDC', savingsAsset: 'sDAI', mode: 'withdraw' },
-    ])
-  })
-
-  test('displays transaction overview', async () => {
-    await withdrawDialog.expectNativeRouteTransactionOverview({
-      routeItems: [
-        {
-          tokenAmount: '10,000.00 sDAI',
-          tokenUsdValue: '$11,255.99',
-        },
-        {
-          tokenAmount: '11,255.99 DAI',
-          tokenUsdValue: '$11,255.99',
-        },
-        {
-          tokenAmount: '11,255.99 USDC',
-          tokenUsdValue: '$11,255.99',
-        },
-      ],
-      outcome: '11,255.99 USDC',
-      outcomeUsd: '$11,255.99',
-    })
-
-    await withdrawDialog.expectUpgradeSwitchToBeHidden()
-  })
-
-  test('executes withdraw', async () => {
-    await withdrawDialog.actionsContainer.acceptAllActionsAction(2)
-
-    await withdrawDialog.expectSuccessPage()
-    await withdrawDialog.clickBackToSavingsButton()
-
-    await savingsPage.expectSupportedStablecoinBalance('USDC', '11,255.99')
   })
 })
